+++ shuup/admin/modules/categories/views/list.py
@@ -9,7 +9,7 @@ from __future__ import unicode_literals
 
 from django.utils.translation import ugettext_lazy as _
 
-from shuup.admin.utils.picotable import ChoicesFilter, Column, TextFilter
+from shuup.admin.utils.picotable import ChoicesFilter, Column, MPTTFilter
 from shuup.admin.utils.views import PicotableListView
 from shuup.core.models import Category, CategoryStatus, CategoryVisibility
 
@@ -18,23 +18,38 @@ class CategoryListView(PicotableListView):
     model = Category
     columns = [
         Column(
-            "name", _(u"Name"), sort_field="translations__name", display="name", linked=True,
-            filter_config=TextFilter(
-                filter_field="translations__name",
-                placeholder=_("Filter by name...")
+            "name", _(u"Name"), sortable=False, display="format_name", linked=True,
+            filter_config=MPTTFilter(
+                choices="get_name_filter_choices",
+                filter_field="id"
             )
         ),
-        Column("status", _(u"Status"), filter_config=ChoicesFilter(choices=CategoryStatus.choices)),
+        Column(
+            "status", _(u"Status"),
+            filter_config=ChoicesFilter(
+               choices=CategoryStatus.choices,
+               default=CategoryStatus.VISIBLE.value
+            )
+        ),
         Column("visibility", _(u"Visibility"), filter_config=ChoicesFilter(choices=CategoryVisibility.choices)),
-        Column("parent", _(u"Parent"), sortable=False, display="parent"),
     ]
+
+    def get_name_filter_choices(self):
+        choices = []
+        for c in Category.objects.all_except_deleted():
+            name = self.format_name(c)
+            choices.append((c.pk, name))
+        return choices
 
     def get_queryset(self):
         return Category.objects.all_except_deleted()
+
+    def format_name(self, instance, *args, **kwargs):
+        level = getattr(instance, instance._mptt_meta.level_attr)
+        return ('---' * level) + ' ' + instance.name
 
     def get_object_abstract(self, instance, item):
         return [
             {"text": "%s" % instance, "class": "header"},
             {"title": _(u"Status"), "text": item["status"]},
-            {"title": _(u"Parent"), "text": item["parent"]} if instance.parent_id else None
         ]